backend/requirements.txt
flask==3.0.3
flask-cors==5.0.0
flask-jwt-extended==4.6.0
flask-socketio==5.3.7
pymongo==4.9.1
python-dotenv==1.0.1
cloudinary==1.41.0
celery==5.4.0
redis==5.1.1
google-generativeai==0.8.3
transformers==4.45.2
torch==2.4.1
Pillow==10.4.0
opencv-python-headless==4.10.0.84
pytesseract==0.3.13
numpy==1.26.4
requests==2.32.3
werkzeug==3.0.4
eventlet==0.37.0
